Usuario

+++ src/main/java/br/edu/ifpb/loteriapweb/repository/ApostaRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import br.edu.ifpb.loteriapweb.enums.StatusAposta;
 import br.edu.ifpb.loteriapweb.model.Aposta;
 
-public interface ApostaRepository extends JpaRepository<Aposta, Long> {
+public interface ApostaRepository extends JpaRepository<Aposta, Integer> {
 
 	List<Aposta> findByStatus(StatusAposta status);
